Update CREATE_TREE.java
This Java program provides a basic implementation of a binary tree data structure. The binary tree is constructed using an array representation, where each element of the array corresponds to a node in the tree. A special value, -1, is used to represent a null node, indicating the absence of a child node.