Update TRAVERSE.java
This Java program defines a basic implementation of a binary tree with three different traversal methods: preorder, inorder, and postorder. Here's a more detailed description:

preorder(node root):

A method to perform preorder traversal (root, left, right) of the binary tree.
Prints the data of each node as it traverses.

inorder(node root):

A method to perform inorder traversal (left, root, right) of the binary tree.
Prints the data of each node as it traverses.

postorder(node root):

A method to perform postorder traversal (left, right, root) of the binary tree.
Prints the data of each node as it traverses.