Create TOP_VIEW.java
This Java program implements a binary tree and a method to print its top view. Here's an extended description:

Given the array {6, 2, -1, 3, -1, 4, -1, -1, 1, -1, 5, -1, -1}, the constructed binary tree would look like this:
       6
      / \
     2   1
      \   \
       3   5
        \
         4
The top view of this tree would be 2 6 1 5

The top view of a binary tree is the set of nodes visible when the tree is viewed from the top. This is determined by considering the first node encountered at each horizontal distance from the root during a level-order traversal.