Create SUBTREE.java
his Java class Solution provides a method to determine whether a binary tree (subRoot) is a subtree of another binary tree (root).
If root is a binary tree and subRoot is a smaller tree, this method will return true if subRoot is a subtree of root. Otherwise, it returns false.
Time Complexity:
The time complexity of this approach is 
𝑂(n*m)
O(n*m), where 
n is the number of nodes in root and 
m is the number of nodes in subRoot. This is because, in the worst case, for every node in root, the algorithm checks whether the subRoot tree matches the subtree rooted at that node.