Create MERGE_SORTED_ARRAY.java
This is a solution to the LeetCode problem "Merge Sorted Array" in Java. The problem asks to merge two sorted integer arrays nums1 and nums2 into one sorted array. The number of elements initialized in nums1 and nums2 are m and n respectively. The final sorted array should not be returned by the function, but instead be stored inside the array nums1. To accommodate this, nums1 has a length of m + n, where the first m elements denote the elements that should be merged, and the last n elements are set to 0 and should be ignored. nums2 has a length of n.