Create MISSING _NUM.java
The given class Solution contains a method missingNumber that calculates the missing number in a sequence from 1 to n when an array of size n-1 is provided. The array contains all the numbers from 1 to n except one missing number.

Time Complexity:

The time complexity of this method is 
0(n)
O(n) due to the two loops that run through the array and the natural numbers.
Space Complexity:

The space complexity is 
0(1)
O(1) as only a few integer variables are used to store sums.